Change score column datatype (int to double)

+++ src/main/java/org/gooru/nucleus/handlers/dataclass/api/processors/repositories/activejdbc/dbhandlers/StudentCoursePerfHandler.java
@@ -146,9 +146,9 @@ class StudentCoursePerfHandler implements DBHandler {
               completedCountMap.forEach(scoreCompletonMap -> {
                 unitData.put(AJEntityBaseReports.ATTR_COMPLETED_COUNT,
                         Integer.valueOf(scoreCompletonMap.get(AJEntityBaseReports.ATTR_COMPLETED_COUNT).toString()));
-                unitData.put(AJEntityBaseReports.ATTR_SCORE, Integer.valueOf(scoreCompletonMap.get(AJEntityBaseReports.ATTR_SCORE).toString()));
+                unitData.put(AJEntityBaseReports.ATTR_SCORE, Double.valueOf(scoreCompletonMap.get(AJEntityBaseReports.ATTR_SCORE).toString()));
                 LOGGER.debug("UnitID : {} - UserID : {} - Score : {}", unitId, userID,
-                        Integer.valueOf(scoreCompletonMap.get(AJEntityBaseReports.ATTR_SCORE).toString()));
+                        Double.valueOf(scoreCompletonMap.get(AJEntityBaseReports.ATTR_SCORE).toString()));
                 LOGGER.debug("UnitID : {} - UserID : {} - completedCount : {}", unitId, userID,
                         Integer.valueOf(scoreCompletonMap.get(AJEntityBaseReports.ATTR_COMPLETED_COUNT).toString()));
